requirements.txt
pytest
scikit-image
lightning
tensorboard
autopep8
PyMCubes
trimesh
pycollada
matplotlib
torchmetrics[image]
mlflow